WebApp/requirements.txt
python-dotenv
fastapi
azure-cosmos
openai
streamlit
python-docx
azure-search-documents 
markitdown
PyPDF2
numpy
